renamed to better fit MVC naming conventions
