+++ attributesModel.py
