+++ interfaceGUIView.py
@@ -15,11 +15,11 @@ import tkinter as tk
 from tkinter import ttk
 
 from classModel import *
-from relationships import *
-from saveLoad import *
-from attributes import *
-from interface import *
-from parameters import *
+from relationshipsModel import *
+from saveLoadModel import *
+from attributesModel import *
+from interfaceView import *
+from parametersModel import *
 
 
 # =================================================================================================================================================================================================================================================================================
